Safari-609.1.11.1/LayoutTests/js/dfg-int-overflow-in-loop-expected.txt

Tests that overflowing an integer in a loop and then only using it in an integer context produces a result that complies with double arithmetic.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(0) is -4094336
PASS successfullyParsed is true

TEST COMPLETE
